feat: dashboard

+++ api-backend/src/main/java/fatec/api/pixel/horaextra/dto/DadosRetornoDashboard.java
@@ -3,7 +3,7 @@ package fatec.api.pixel.horaextra.dto;
 public record DadosRetornoDashboard(double horas,
 									String razaoSocial,
 									String nome,
-									Long idUsuario,
+									int idUsuario,
 									String Modalidade) {
 
 }
